Get authenticated user and authorization

+++ backend/api/v1/route_jobs.py
@@ -4,13 +4,14 @@ from sqlalchemy.orm import Session
 from db.session import get_db
 from schemas.jobs import ShowJob, JobSchema
 from repository.jobs import create_new_job, find_job_by_id, get_all_jobs, update_a_job, delete_a_job
+from repository.login import get_current_user
+from models.users import User
 
 router = APIRouter()
 
 @router.post('/', response_model=ShowJob)
-def create_job(job: JobSchema, db: Session = Depends(get_db)):
-    owner_id = 1
-    job = create_new_job(job, owner_id, db)
+def create_job(job: JobSchema, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
+    job = create_new_job(job=job, owner_id=current_user.id, db=db)
     return job
 
 
@@ -34,15 +35,19 @@ def show_job(id: int, db: Session = Depends(get_db)):
     return job
 
 @router.put('/{id}')
-def update_job(id: int, job: JobSchema, db: Session = Depends(get_db)):
-    
-    owner_id = 1
+def update_job(id: int, job: JobSchema, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
 
     job_found = find_job_by_id(id, db)
     if not job_found:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Job with index {id} not exist"
+        )
+    
+    if job_found.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=f"Your are not autorise to update this job post"
         )
     
     update_a_job(job_found, job, db)
@@ -52,7 +57,7 @@ def update_job(id: int, job: JobSchema, db: Session = Depends(get_db)):
     }
 
 @router.delete('/{id}')
-def delete_job(id: int, db: Session = Depends(get_db)):
+def delete_job(id: int, db: Session = Depends(get_db), current_user: User = Depends(get_current_user)):
     job_found = find_job_by_id(id, db)
 
     if not job_found:
@@ -60,9 +65,15 @@ def delete_job(id: int, db: Session = Depends(get_db)):
             status_code=status.HTTP_404_NOT_FOUND,
             detail=f"Job with index {id} not exist"
         )
+    
+    if job_found.owner_id != current_user.id:
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail=f"Your are not autorise to update this job post"
+        )
 
     delete_a_job(job_found, db)
     return {
         "msg": "job deleted with success",
-        "status": status.HTTP_200_OK,
+        "status": status.HTTP_200_OK
     }
